Added a boilerplate and styles around header
Started the project by adding a boilerplate and inputting as much as
content inside the html code by also implementing semantic html by using
nav and section since there are 4 sections for this project.

Completed the header section of the project by using margin left to
make sure that the links are on the right-hand side of the web and by
using a container to ensure that all sections including header has the
same left and right space to achieve the desired outcome.